* search: (25 commits)
  Blackout for new search, forcing non-staff to continue using old search.
  Adding checks for whether or not a feed is indexed before indexing new stories. Also handling story removal properly.
  Handling up to 10000 feed updates per 5 minutes.
  Fixing issue with scheduling feeds.
  Showing indexing tooltip on hover. Also correctly hiding tooltip.
  Adding remove search index method. Also removing source on both autocomplete and story search index to save space.
  Preserving search during oldest/newest change.
  Fixing size issue with elasticsearch.
  Fixing bug in paging queries.
  Preserving order on searches. Also using elasticsearch's paging.
  Fixing bug with not correctly blurring search.
  Showing correct unread status in river/unread during search.
  Lightening feedback colors.
  Retrying search on search indexing completion. Hiding story content preview when too short.
  Adding migration for creating search mapping.
  Updating search indexing progress in real-time.
  Adding task to index user's subscriptions and alert them to progress.
  Indexing stories on indexes feeds during feed fetch.
  Searching read feeds as well.
  Default null for new search indexed field for backwards compatibility.
  ...

+++ apps/search/models.py
@@ -1,8 +1,125 @@
+import time
+import datetime
 import pyes
-from pyes.query import FuzzyQuery, MatchQuery, PrefixQuery
+import redis
+import mongoengine as mongo
+from pyes.query import MatchQuery
 from django.conf import settings
 from django.contrib.auth.models import User
+from apps.search.tasks import IndexSubscriptionsForSearch
 from utils import log as logging
+
+class MUserSearch(mongo.Document):
+    '''Search index state of a user's subscriptions.'''
+    user_id                  = mongo.IntField(unique=True)
+    last_search_date         = mongo.DateTimeField()
+    subscriptions_indexed    = mongo.BooleanField()
+    subscriptions_indexing   = mongo.BooleanField()
+    
+    meta = {
+        'collection': 'user_search',
+        'indexes': ['user_id'],
+        'index_drop_dups': True,
+        'allow_inheritance': False,
+    }
+    
+    @classmethod
+    def get_user(cls, user_id):
+        try:
+            user_search = cls.objects.get(user_id=user_id)
+        except cls.DoesNotExist:
+            user_search = cls.objects.create(user_id=user_id)
+        
+        return user_search
+    
+    def touch_search_date(self):
+        # Blackout
+        # if not self.subscriptions_indexed and not self.subscriptions_indexing:
+        #     self.schedule_index_subscriptions_for_search()
+        #     self.subscriptions_indexing = True
+
+        self.last_search_date = datetime.datetime.now()
+        self.save()
+
+    def schedule_index_subscriptions_for_search(self):
+        IndexSubscriptionsForSearch.apply_async(kwargs=dict(user_id=self.user_id))
+        
+    # Should be run as a background task
+    def index_subscriptions_for_search(self):
+        from apps.rss_feeds.models import Feed
+        from apps.reader.models import UserSubscription
+        
+        SearchStory.create_elasticsearch_mapping()
+        
+        start = time.time()
+        not_found = 0
+        processed = 0
+        user = User.objects.get(pk=self.user_id)
+        r = redis.Redis(connection_pool=settings.REDIS_PUBSUB_POOL)
+        r.publish(user.username, 'search_index_complete:start')
+        throttle = time.time()
+        
+        subscriptions = UserSubscription.objects.filter(user=user)
+        total = subscriptions.count()
+        logging.user(user, "~FCIndexing ~SB%s feeds~SN for ~SB~FB%s~FC~SN..." % 
+                     (total, user.username))
+        
+        for sub in subscriptions:
+            try:
+                feed = sub.feed
+            except Feed.DoesNotExist:
+                not_found += 1
+                continue
+            
+            feed.index_stories_for_search()
+            processed += 1
+            
+            # Throttle notifications to client as to not flood them
+            if time.time() - throttle > 0.05:
+                r.publish(user.username, 'search_index_complete:%.4s' % (float(processed)/total))
+                throttle = time.time()
+        
+        duration = time.time() - start
+        logging.user(user, "~FCIndexed ~SB%s/%s feeds~SN for ~SB~FB%s~FC~SN in ~FM~SB%s~FC~SN sec." % 
+                     (processed, total, user.username, round(duration, 2)))
+        r.publish(user.username, 'search_index_complete:done')
+        
+        self.subscriptions_indexed = True
+        self.subscriptions_indexing = False
+        self.save()
+    
+    @classmethod
+    def remove_all(cls):
+        user_searches = cls.objects.all()
+        logging.info(" ---> ~SN~FRRemoving ~SB%s~SN user searches..." % user_searches.count())
+        for user_search in user_searches:
+            user_search.remove()
+            
+        logging.info(" ---> ~FRRemoving stories search index...")
+        SearchStory.drop()
+        
+    def remove(self):
+        from apps.rss_feeds.models import Feed
+        from apps.reader.models import UserSubscription
+
+        user = User.objects.get(pk=self.user_id)
+        subscriptions = UserSubscription.objects.filter(user=self.user_id, 
+                                                        feed__search_indexed=True)
+        total = subscriptions.count()
+        removed = 0
+        
+        for sub in subscriptions:
+            try:
+                feed = sub.feed
+            except Feed.DoesNotExist:
+                continue
+            feed.search_indexed = False
+            feed.save()
+            removed += 1
+            
+        logging.user(user, "~FCRemoved ~SB%s/%s feed's search indexes~SN for ~SB~FB%s~FC~SN." % 
+                     (removed, total, user.username))
+        self.delete()
 
 class SearchStory:
     
@@ -18,88 +135,81 @@ class SearchStory:
         return "%s-type" % cls.name
         
     @classmethod
-    def create_elasticsearch_mapping(cls):
-        cls.ES.create_index("%s-index" % cls.name)
+    def create_elasticsearch_mapping(cls, delete=False):
+        if delete:
+            cls.ES.indices.delete_index_if_exists("%s-index" % cls.name)
+        cls.ES.indices.create_index_if_missing("%s-index" % cls.name)
         mapping = { 
             'title': {
-                'boost': 2.0,
+                'boost': 3.0,
                 'index': 'analyzed',
-                'store': 'yes',
+                'store': 'no',
                 'type': 'string',
-                "term_vector" : "with_positions_offsets"
+                'analyzer': 'snowball',
             },
             'content': {
                 'boost': 1.0,
                 'index': 'analyzed',
-                'store': 'yes',
+                'store': 'no',
                 'type': 'string',
-                "term_vector" : "with_positions_offsets"
+                'analyzer': 'snowball',
             },
             'author': {
                 'boost': 1.0,
                 'index': 'analyzed',
-                'store': 'yes',
+                'store': 'no',
                 'type': 'string',   
-            },
-            'db_id': {
-                'index': 'not_analyzed',
-                'store': 'yes',
-                'type': 'string',   
+                'analyzer': 'keyword',
             },
             'feed_id': {
-                'store': 'yes',
+                'store': 'no',
                 'type': 'integer'
             },
             'date': {
-                'store': 'yes',
+                'store': 'no',
                 'type': 'date',
-            },
-            'user_ids': {
-                'index': 'not_analyzed',
-                'store': 'yes',
-                'type': 'integer',
-                'index_name': 'user_id'
             }
         }
-        cls.ES.put_mapping("%s-type" % cls.name, {'properties': mapping}, ["%s-index" % cls.name])
+        cls.ES.indices.put_mapping("%s-type" % cls.name, {
+            'properties': mapping,
+            '_source': {'enabled': False},
+        }, ["%s-index" % cls.name])
         
     @classmethod
-    def index(cls, user_id, story_id, story_title, story_content, story_author, story_date, db_id):
+    def index(cls, story_hash, story_title, story_content, story_author, story_feed_id, 
+              story_date):
         doc = {
-            "content": story_content,
-            "title": story_title,
-            "author": story_author,
-            "date": story_date,
-            "user_ids": user_id,
-            "db_id": db_id,
+            "content"   : story_content,
+            "title"     : story_title,
+            "author"    : story_author,
+            "feed_id"   : story_feed_id,
+            "date"      : story_date,
         }
-        cls.ES.index(doc, "%s-index" % cls.name, "%s-type" % cls.name, story_id)
+        cls.ES.index(doc, "%s-index" % cls.name, "%s-type" % cls.name, story_hash)
+    
+    @classmethod
+    def remove(cls, story_hash):
+        cls.ES.delete("%s-index" % cls.name, "%s-type" % cls.name, story_hash)
         
     @classmethod
-    def query(cls, user_id, text):
-        user = User.objects.get(pk=user_id)
-        cls.ES.refresh()
-        q = pyes.query.StringQuery(text)
-        print q.serialized(), cls.index_name, cls.type_name
-        results = cls.ES.search(q, indices=cls.index_name, doc_types=[cls.type_name])
-        logging.user(user, "~FGSearch ~FCstories~FG for: ~SB%s" % text)
+    def drop(cls):
+        cls.ES.delete("%s-index" % cls.name, "%s-type" % cls.name, None)
         
-        if not results.total:
-            logging.user(user, "~FGSearch ~FCstories~FG by title: ~SB%s" % text)
-            q = FuzzyQuery('title', text)
-            results = cls.ES.search(q)
-            
-        if not results.total:
-            logging.user(user, "~FGSearch ~FCstories~FG by content: ~SB%s" % text)
-            q = FuzzyQuery('content', text)
-            results = cls.ES.search(q)
-            
-        if not results.total:
-            logging.user(user, "~FGSearch ~FCstories~FG by author: ~SB%s" % text)
-            q = FuzzyQuery('author', text)
-            results = cls.ES.search(q)
-            
-        return results
+    @classmethod
+    def query(cls, feed_ids, query, order, offset, limit):
+        cls.create_elasticsearch_mapping()
+        cls.ES.indices.refresh()
+        
+        sort     = "date:desc" if order == "newest" else "date:asc"
+        string_q = pyes.query.StringQuery(query, default_operator="AND")
+        feed_q   = pyes.query.TermsQuery('feed_id', feed_ids)
+        q        = pyes.query.BoolQuery(must=[string_q, feed_q])
+        results  = cls.ES.search(q, indices=cls.index_name(), doc_types=[cls.type_name()],
+                                 partial_fields={}, sort=sort, start=offset, size=limit)
+        logging.info(" ---> ~FG~SNSearch ~FCstories~FG for: ~SB%s~SN (across %s feed%s)" % 
+                     (query, len(feed_ids), 's' if len(feed_ids) != 1 else ''))
+        
+        return [r.get_id() for r in results]
 
 
 class SearchFeed:
@@ -116,8 +226,9 @@ class SearchFeed:
         return "%s-type" % cls.name
         
     @classmethod
-    def create_elasticsearch_mapping(cls):
-        cls.ES.indices.delete_index_if_exists("%s-index" % cls.name)
+    def create_elasticsearch_mapping(cls, delete=False):
+        if delete:
+            cls.ES.indices.delete_index_if_exists("%s-index" % cls.name)
         settings =  {
             "index" : {
                 "analysis": {
@@ -161,7 +272,7 @@ class SearchFeed:
                 }
             }
         }
-        cls.ES.indices.create_index("%s-index" % cls.name, settings)
+        cls.ES.indices.create_index_if_missing("%s-index" % cls.name, settings)
 
         mapping = {
             "address": {
@@ -186,16 +297,19 @@ class SearchFeed:
                 "type": "string"
             }
         }
-        cls.ES.indices.put_mapping("%s-type" % cls.name, {'properties': mapping}, ["%s-index" % cls.name])
+        cls.ES.indices.put_mapping("%s-type" % cls.name, {
+            'properties': mapping,
+            '_source': {'enabled': False},
+        }, ["%s-index" % cls.name])
         
     @classmethod
     def index(cls, feed_id, title, address, link, num_subscribers):
         doc = {
-            "feed_id": feed_id,
-            "title": title,
-            "address": address,
-            "link": link,
-            "num_subscribers": num_subscribers,
+            "feed_id"           : feed_id,
+            "title"             : title,
+            "address"           : address,
+            "link"              : link,
+            "num_subscribers"   : num_subscribers,
         }
         cls.ES.index(doc, "%s-index" % cls.name, "%s-type" % cls.name, feed_id)
         
@@ -206,21 +320,18 @@ class SearchFeed:
         
         logging.info("~FGSearch ~FCfeeds~FG by address: ~SB%s" % text)
         q = MatchQuery('address', text, operator="and", type="phrase")
-        print q.serialize(), cls.index_name(), cls.type_name()
         results = cls.ES.search(query=q, sort="num_subscribers:desc", size=5,
                                 doc_types=[cls.type_name()])
 
         if not results.total:
             logging.info("~FGSearch ~FCfeeds~FG by title: ~SB%s" % text)
             q = MatchQuery('title', text, operator="and")
-            print q.serialize()
             results = cls.ES.search(query=q, sort="num_subscribers:desc", size=5,
                                     doc_types=[cls.type_name()])
             
         if not results.total:
             logging.info("~FGSearch ~FCfeeds~FG by link: ~SB%s" % text)
             q = MatchQuery('link', text, operator="and")
-            print q.serialize()
             results = cls.ES.search(query=q, sort="num_subscribers:desc", size=5,
                                     doc_types=[cls.type_name()])
